[AD-602] Split DocumentDbTableSchemaGeneratorTest into 3 categories
